requirements.txt
# Core dependencies
Flask==2.3.3
Werkzeug==2.3.7
flask-apispec==0.11.4
Flask-Cors==3.0.10
Flask-RESTful==0.3.10
Flask-SocketIO==5.5.1

# Web socket and async
# Use binary wheel for gevent (avoid compilation)
gevent==22.10.2; platform_system != "Darwin" or platform_machine != "arm64"
gevent==22.10.2; platform_system == "Darwin" and platform_machine == "arm64" --no-binary :all:
gevent-websocket==0.10.1
python-socketio>=5.12.0
pyzmq==25.1.1
socketio-client==0.7.2

# AI/ML dependencies
openfabric-pysdk==0.2.9
runstats
transformers>=4.43.0
torch>=2.0.0
huggingface_hub>=0.16.0
accelerate>=0.21.0

# Memory and database
chromadb>=0.4.0
sentence-transformers>=2.2.0
sqlalchemy>=2.0.0

# API and utilities
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
marshmallow>=3.20.0
marshmallow-enum==1.5.1
marshmallow-jsonapi==0.24.0
marshmallow-jsonschema==0.13.0
deepdiff>=7.0.1
pickleDB==0.9.2
web3>=7.5.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0
tqdm>=4.62.3

# UI
gradio>=3.40.0
pygltflib
